# doc/api.yaml
openapi: 3.0.0
info:
  title: Chat Application API
  description: Chat application REST API documentation
  version: 1.0.0
  contact:
    name: Chat App Team
    email: support@chatapp.com
    url: https://chatapp.com/support
servers:
- url: /api/v1
  description: API foundation path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT authentication using the Bearer scheme.
        Example: "Authorization: Bearer {token}"
  responses:
    BadRequest:
      description: >
        The request was malformed or contained invalid parameters.
        Check the error details for specific validation issues.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: "Invalid request parameters"
            trace: "req-123456"
            error: "validation.failed"
            errors:
              - "username: Must be between 3-50 characters"
              - "email: Invalid email format"
    Unauthorized:
      description: >
        Authentication failed or user lacks proper permissions.
        Include a valid JWT token in the Authorization header.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            message: "Unauthorized: Invalid or expired token"
            trace: "req-789012"
            error: "auth.invalid_token"
            errors: []      
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: "User not found"
            trace: "req-345678"
            error: "resource.not_found"
            errors: []     
    Forbidden:
      description: >
        Authenticated user does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    FileUpload:
      type: object
      description: File upload information model
      required:
      - filename
      - url
      properties:
        filename:
          type: string
          description: Original filename
          minLength: 1
          maxLength: 255
          pattern: ^[a-zA-Z0-9_\-. ]+$
        size:
          type: integer
          format: int64
          description: File size in bytes
          minimum: 1
          maximum: 10485760  # 10MB
        url:
          type: string
          description: File access URL
          format: url
          minLength: 1
          maxLength: 255
          pattern: ^https?://[^\s/$.?#].[^\s]*$
    FileUploadEnvelope:
      description: Metadata and status returned after a file upload completes.
      allOf:
        - $ref: '#/components/schemas/BaseSuccessResponse'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              description: Payload containing uploaded file metadata
              required:
                - file
              properties:
                file:
                  $ref: '#/components/schemas/FileUpload'
...
    Error:
      type: object
      description: Standard error response model for API error conditions
      required:
      - code
      - message
      - trace
      - error
      - errors
      properties:
        code:
          type: integer
          format: int64
          description: HTTP error code
          minimum: 400
          maximum: 599
        message:
          type: string
          description: Human-readable error message
          minLength: 1
          maxLength: 255
          pattern: ^[\s\S]{1,255}$
        trace:
          type: string
          description: Error trace ID for debugging purposes
          pattern: ^[a-zA-Z0-9\s.,!?-]+$
          minLength: 1
          maxLength: 1000
        error:
          type: string
          description: Machine-readable error key (stable identifier for programmatic handling)
          minLength: 1
          maxLength: 100
          pattern: ^[a-z0-9_.:-]+$
          example: auth.invalid_token
        errors:
          type: array
          description: Collection of detailed error messages
          items:
            type: string
            minLength: 1
            maxLength: 255
            pattern: ^[\s\S]{1,255}$
          minItems: 0
          maxItems: 100
    BaseSuccessResponse:
      type: object
      description:  Base structure for all success responses
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int64
          description: HTTP status code indicating success
          enum:
          - 200
          - 201
          minimum: 200
          maximum: 201
          example: 200
        message:
          type: string
          description: Human-readable success message
          minLength: 1
          maxLength: 255
          pattern: ^[\s\S]{1,1000}$
          example: "Operation succeeded"  
...
    Group:
      type: object
      description: Group information model containing group details and members
      required:
      - id
      - name
      - conversation_id
      - members
      properties:
        id:
          type: string
          description: Unique identifier for the group
          minLength: 1
          maxLength: 64
          pattern: ^[a-zA-Z0-9-]+$
        name:
          type: string
          description: Group display name
          minLength: 1
          maxLength: 100
          pattern: ^[a-zA-Z0-9\s'-]+$
        conversation_id:
          type: string
          description: ID of the present conversation
          pattern: ^[a-zA-Z0-9-]+$
          minLength: 1
          maxLength: 64
        photo_url:
          type: string
          format: url
          description: URL of group photo
          minLength: 1
          maxLength: 255
          pattern: ^https?://[^\s/$.?#].[^\s]*$
        members:
          type: array
          description: Array of group member objects
          minItems: 1
          maxItems: 100
          items:
            type: object
            description: Group member information
            required:
            - user_id
            - role
            properties:
              user_id:
                type: string
                description: User ID of the member
                minLength: 1
                maxLength: 64
                pattern: ^[a-zA-Z0-9-]+$
              role:
                type: string
                description: Member's role in the group
                enum:
                - admin
                - member
                minLength: 1
                maxLength: 50
    GroupEnvelope:
      description: Envelope containing a single group resource.
      allOf:
        - $ref: '#/components/schemas/BaseSuccessResponse'
        - type: object
          required: 
            - data
          properties:
            data:
              type: object
              description: Payload containing a single group resource
              required: 
                - group
              properties:
                group:
                  $ref: '#/components/schemas/Group'
    GroupEnvelopeCollection:
      description: Envelope containing a list of groups and optional paging info.
      allOf:
        - $ref: '#/components/schemas/BaseSuccessResponse'
        - type: object
          description: Success response with a list of group resources
          required: 
            - items
          properties:
            items:
              type: array
              description: List of groups
              minItems: 0
              maxItems: 100
              items:
                $ref: '#/components/schemas/Group'
    AddGroupMembersRequest:
      type: object
      description: Request model for adding members to an existing group
      required:
      - member_ids
      properties:
        member_ids:
          type: array
          description: Array of user IDs to add to the group
          minItems: 1
          maxItems: 100
          items:
            type: string
            description: User ID to add to the group
            pattern: ^[a-zA-Z0-9-]+$
            minLength: 1
            maxLength: 64
...
  parameters:
...
    id:
      name: id
      in: path
      required: true
      description: ID of the resource
      schema:
        type: string
        pattern: ^[a-zA-Z0-9-]+$
        minLength: 1
        maxLength: 64
        description: uniquie identifier
paths:
...
  /users/set_photo:
    put:
      tags:
      - users
      operationId: setMyPhoto
      summary: Change profile photo
      description: Update the current user's profile picture
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - upload
              description: Request to update profile photo
              properties:
                upload:
                  type: string
                  format: binary
                  description:  Image file (JPEG/PNG)
                  minLength: 1
                  maxLength: 10485760
      responses:
        '200':
          description: Profile photo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /groups:
    post:
      tags:
      - groups
      operationId: createGroup
      summary: Create group
      description: Create group and associated conversation
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Create a new group 
              required: 
              - name
              - member_ids
              properties:
                name:
                  type: string
                  description: Group display name 
                  minLength: 1
                  maxLength: 100
                  pattern: "^[a-zA-Z0-9\\s'-]+$"
                member_ids:
                  type: array
                  description: IDs of users to include
                  minItems: 1
                  maxItems: 100
                  items:
                    type: string
                    description: User ID to add to the group
                    pattern: '^[a-zA-Z0-9-]+$'
                    minLength: 1
                    maxLength: 64
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupEnvelope'
              example:
                code: 201
                message: Group created
                data:
                  group:
                    id: '1'
                    name: Group 1
                    conversation_id: 'c-1'  
                    members:
                      - user_id: '1'
                        role: admin
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags:
      - groups
      operationId: getGroupsList
      summary: Get groups list
      description: Retrieve list of user's groups
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Groups list retrieved successfully
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/GroupEnvelopeCollection'
              example:
                code: 200
                message: Groups list retrieved
                items:                        
                  - id: '1'
                    name: Group 1
                    conversation_id: 'c-1'        
                    members:
                      - user_id: '1'
                        role: admin
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags:
      - groups
      operationId: getGroupDetail
      summary: Get group details
      description: Retrieve details of a specific group
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Group details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupEnvelope'
              example:
                code: 200
                message: Group details retrieved
                data:
                    group:
                      id: '1'
                      name: Group 1
                      conversation_id: 'c-1' 
                      members:
                        - user_id: '1'
                          role: admin
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /groups/{id}/photo:
    parameters:
      - $ref: '#/components/parameters/id'
    put:
      tags:
      - groups
      operationId: setGroupPhoto
      summary: Change group photo
      description: Update the group's profile picture
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Request to update group photo
              required:
                - upload
              properties:
                upload:
                  type: string
                  format: binary
                  description: Group photo file
                  minLength: 1
                  maxLength: 10485760
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/members:
    parameters:
      - $ref: '#/components/parameters/id'
    post:
      tags:
      - groups
      operationId: addToGroup
      summary: Add group members
      description: Add new members to an existing group
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroupMembersRequest'
      responses:
        '200':
          description: Members added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
              example:
                code: 200
                message: Members added
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - groups
      operationId: leaveGroup
      summary: Leave group
      description: Current user leaves the specified group
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Successfully left the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
              example:
                code: 200
                message: Left the group
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...